+++ tests/test_differentiable.py
@@ -14,13 +14,13 @@ def test_simple_quadrupole():
     """
     segment = cheetah.Segment(
         [
-            cheetah.Drift(length=torch.tensor(1.0)),
+            cheetah.Drift(length=1.0),
             cheetah.Quadrupole(
-                length=torch.tensor(0.2),
+                length=0.2,
                 k1=nn.Parameter(torch.tensor(3.142)),
                 name="my_quad",
             ),
-            cheetah.Drift(length=torch.tensor(1.0)),
+            cheetah.Drift(length=1.0),
         ]
     )
     incoming_beam = cheetah.ParticleBeam.from_astra(
@@ -120,7 +120,7 @@ def test_nonleaf_tracking(ElementClass):
     segment = cheetah.Segment(
         elements=[
             cheetah.Drift(length=torch.tensor(1.0, requires_grad=True)),
-            ElementClass(length=torch.tensor(2.0)),
+            ElementClass(length=2.0),
         ]
     )
     segment.track(beam)
@@ -154,7 +154,7 @@ def test_parameters_at_initialization():
     Test that passing a `torch.nn.Parameter` at initialization registeres the parameter
     in the same way as an assignment after initialization.
     """
-    dipole_with_buffer = cheetah.Dipole(length=torch.tensor(1.0))
+    dipole_with_buffer = cheetah.Dipole(length=1.0)
 
     # Dipole with buffer (without parameter) should not have any parameters
     assert len(list(dipole_with_buffer.parameters())) == 0
@@ -162,8 +162,8 @@ def test_parameters_at_initialization():
     # Create two dipoles with the same parameter, one passed at initialization and one
     # assigned after initialization.
     parameter = torch.nn.Parameter(torch.tensor(0.2))
-    dipole_initial = cheetah.Dipole(length=torch.tensor(1.0), angle=parameter)
-    dipole_assigned = cheetah.Dipole(length=torch.tensor(1.0))
+    dipole_initial = cheetah.Dipole(length=1.0, angle=parameter)
+    dipole_assigned = cheetah.Dipole(length=1.0)
     dipole_assigned.angle = parameter
 
     # Both dipoles should have the same parameter (the originally passed one and one in